backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic
pydantic-settings
email-validator
passlib[bcrypt]
python-jose[cryptography]
python-multipart
requests
pytest
pytest-asyncio
httpx
anthropic
